Changed type of Longitude and Latitude in CityLocation

+++ SpotFinder/SpotFinder/Models/Core/CityLocation.cs
@@ -4,7 +4,7 @@ namespace SpotFinder.Models.Core
     public class CityLocation
     {
         public string City { get; set; }
-        public int? Longitude { get; set; }
-        public int? Latitude { get; set; }
+        public double? Longitude { get; set; }
+        public double? Latitude { get; set; }
     }
 }
